+++ src/test/java/keeper/project/homepage/systemadmin/service/SystemAdminServiceTest.java
@@ -13,13 +13,12 @@ import java.util.stream.Stream;
 import keeper.project.homepage.systemadmin.dto.response.JobResponseDto;
 import keeper.project.homepage.systemadmin.dto.response.MemberJobTypeResponseDto;
 import keeper.project.homepage.systemadmin.dto.response.TypeResponseDto;
-import keeper.project.homepage.systemadmin.service.SystemAdminService;
 import keeper.project.homepage.util.entity.ThumbnailEntity;
 import keeper.project.homepage.member.entity.MemberEntity;
 import keeper.project.homepage.member.entity.MemberHasMemberJobEntity;
 import keeper.project.homepage.member.entity.MemberJobEntity;
 import keeper.project.homepage.member.entity.MemberTypeEntity;
-import keeper.project.homepage.clerk.exception.CustomClerkInaccessibleJobException;
+import keeper.project.homepage.systemadmin.exception.CustomClerkInaccessibleJobException;
 import keeper.project.homepage.member.repository.MemberHasMemberJobRepository;
 import keeper.project.homepage.member.repository.MemberJobRepository;
 import keeper.project.homepage.member.service.MemberUtilService;
